Module05/requirements.txt
db-dtypes==1.1.1
Flask==2.3.2
Flask-Session==0.5.0
google-cloud-storage==1.42.0
google-cloud-bigquery==3.11.3
gunicorn==20.1.0
pandas==1.5.2
pyarrow==12.0.1